lparser.c: "function" keyword optional after "local".
This bit of syntactic sugar is not backwards compatible. It changes
the meaning of code where a local statement that declares a single
variable without assignment is followed by a function call statement
starting with an opening round bracket. For example,

  local f(a)(b)

Before the change it is parsed as the two statements

  local f
  a(b)

After the chenge it is parsed as

  local function f(a)
    (b)

Code that relies on the old syntax will be rejected as a syntax error
because of the change in nesting.